# protocol/tx: replace a call to io.CopyN with a fixed-length read and write