```
benchmark                     old ns/op     new ns/op     delta
BenchmarkHashEmptyTx-8        13705         5998          -56.23%
BenchmarkHashNonemptyTx-8     56714         28370         -49.98%
```